Split up the SRP group into its own file
